src/main/webapp/app/entities/ad-client/ad-client.component.ts
import { mixins } from 'vue-class-component';

import { Component, Inject } from 'vue-property-decorator';
import Vue2Filters from 'vue2-filters';
import { IADClient } from '@/shared/model/ad-client.model';
import AlertMixin from '@/shared/alert/alert.mixin';

import ADClientService from './ad-client.service';

@Component
export default class ADClient extends mixins(Vue2Filters.mixin, AlertMixin) {
  @Inject('aDClientService') private aDClientService: () => ADClientService;
  private removeId: number = null;
  public itemsPerPage = 20;
  public queryCount: number = null;
  public page = 1;
  public previousPage = 1;
  public propOrder = 'id';
  public reverse = false;
  public totalItems = 0;

  public aDClients: IADClient[] = [];

  public isFetching = false;

  public mounted(): void {
    this.retrieveAllADClients();
  }

  public clear(): void {
    this.page = 1;
    this.retrieveAllADClients();
  }

  public retrieveAllADClients(): void {
    this.isFetching = true;

    const paginationQuery = {
      page: this.page - 1,
      size: this.itemsPerPage,
      sort: this.sort()
    };
    this.aDClientService()
      .retrieve(paginationQuery)
      .then(
        res => {
          this.aDClients = res.data;
          this.totalItems = Number(res.headers['x-total-count']);
          this.queryCount = this.totalItems;
          this.isFetching = false;
        },
        err => {
          this.isFetching = false;
        }
      );
  }

  public prepareRemove(instance: IADClient): void {
    this.removeId = instance.id;
    if (<any>this.$refs.removeEntity) {
      (<any>this.$refs.removeEntity).show();
    }
  }

  public removeADClient(): void {
    this.aDClientService()
      .delete(this.removeId)
      .then(() => {
        const message = this.$t('opusWebApp.aDClient.deleted', { param: this.removeId });
        this.alertService().showAlert(message, 'danger');
        this.getAlertFromStore();
        this.removeId = null;
        this.retrieveAllADClients();
        this.closeDialog();
      });
  }

  public sort(): Array<any> {
    const result = [this.propOrder + ',' + (this.reverse ? 'asc' : 'desc')];
    if (this.propOrder !== 'id') {
      result.push('id');
    }
    return result;
  }

  public loadPage(page: number): void {
    if (page !== this.previousPage) {
      this.previousPage = page;
      this.transition();
    }
  }

  public transition(): void {
    this.retrieveAllADClients();
  }

  public changeOrder(propOrder): void {
    this.propOrder = propOrder;
    this.reverse = !this.reverse;
    this.transition();
  }

  public closeDialog(): void {
    (<any>this.$refs.removeEntity).hide();
  }
}
